Rename: Id를 제외한 Long타입 INTEGER로 변경
Member entity의 column들 중에  id 를 제외한 Long타입을 Integer타입으로 변경하였습니다.
추후 필요시 다시 Long타입으로 변환 예정입니다.

+++ src/main/java/com/codein/domain/Member.java
@@ -1,6 +1,6 @@
 package com.codein.domain;
 
-import com.codein.error.exception.MemberNotLoginException;
+import com.codein.error.exception.member.MemberNotLoginException;
 import com.codein.responsedto.LoginResponseDto;
 import com.codein.responsedto.ProfileResponseDto;
 import jakarta.persistence.*;
@@ -47,7 +47,7 @@ public class Member {
     private LocalDateTime createdAt;
 
     @Setter
-    private Long point;
+    private Integer point;
 
     @OneToMany(cascade = CascadeType.ALL, mappedBy = "member", fetch = FetchType.LAZY)
     private final List<Session> sessions = new ArrayList<>();
@@ -65,7 +65,7 @@ public class Member {
         this.birth = birth;
         this.sex = sex;
         this.createdAt = LocalDateTime.now();
-        this.point = 0L;
+        this.point = 0;
     }
 
     public LoginResponseDto changeMemberResponse() {
